requirements.txt
cassandra-driver
pandas
black